+++ foxes/input/windio/read_farm.py
@@ -32,27 +32,27 @@ def read_turbine_type(wio_trbns, algo_dict, ws_exp_P, ws_exp_ct, verbosity):
 
     """
     tname = wio_trbns.pop("name")
-    if verbosity > 1:
+    if verbosity > 2:
         print("    Reading wio_trbns")
         print("      Name:", tname)
         print("      Contents:", [k for k in wio_trbns.keys()])
 
     # read performance:
     performance = Dict(wio_trbns["performance"], name="performance")
-    if verbosity > 1:
+    if verbosity > 2:
         print("        Reading performance")
         print("          Contents:", [k for k in performance.keys()])
 
     # P, ct data:
     if "power_curve" in performance:
         power_curve = Dict(performance["power_curve"], name="power_curve")
-        if verbosity > 1:
+        if verbosity > 2:
             print("            Reading power_curve")
             print("              Contents:", [k for k in power_curve.keys()])
         P = power_curve["power_values"]
         ws_P = power_curve["power_wind_speeds"]
         ct_curve = Dict(performance["Ct_curve"], name="Ct_values")
-        if verbosity > 1:
+        if verbosity > 2:
             print("            Reading Ct_curve")
             print("              Contents:", [k for k in ct_curve.keys()])
         ct = ct_curve["Ct_values"]
@@ -66,7 +66,7 @@ def read_turbine_type(wio_trbns, algo_dict, ws_exp_P, ws_exp_ct, verbosity):
                 raise ValueError(f"Expecting wind speed exponent 1, 2 or 3, got {wse}")
             return FV.REWS if wse == 1 else (FV.REWS2 if wse == 2 else FV.REWS3)
 
-        if verbosity > 1:
+        if verbosity > 2:
             print(f"            Creating model '{tname}'")
             print(f"              Turbine type class: PCtFromTwo")
         algo_dict["mbook"].turbine_types[tname] = TurbineType.new(
@@ -83,19 +83,19 @@ def read_turbine_type(wio_trbns, algo_dict, ws_exp_P, ws_exp_ct, verbosity):
             var_ws_P=_get_wse_var(ws_exp_P),
             rho=1.225,
         )
-        if verbosity > 1:
+        if verbosity > 2:
             print("               ", algo_dict["mbook"].turbine_types[tname])
 
     # P, ct data:
     elif "Cp_curve" in performance:
         cp_curve = Dict(performance["Cp_curve"], name="Cp_curve")
-        if verbosity > 1:
+        if verbosity > 2:
             print("            Reading Cp_curve")
             print("              Contents:", [k for k in cp_curve.keys()])
         cp = cp_curve["Cp_values"]
         ws_cp = cp_curve["Cp_wind_speeds"]
         ct_curve = Dict(performance["Ct_curve"], name="Ct_values")
-        if verbosity > 1:
+        if verbosity > 2:
             print("            Reading Ct_curve")
             print("              Contents:", [k for k in ct_curve.keys()])
         ct = ct_curve["Ct_values"]
@@ -104,7 +104,7 @@ def read_turbine_type(wio_trbns, algo_dict, ws_exp_P, ws_exp_ct, verbosity):
         data_cp = pd.DataFrame(data={"ws": ws_cp, "cp": cp})
         data_ct = pd.DataFrame(data={"ws": ws_ct, "ct": ct})
 
-        if verbosity > 1:
+        if verbosity > 2:
             print(f"            Creating model '{tname}'")
             print(f"              Turbine type class: CpCtFromTwo")
         algo_dict["mbook"].turbine_types[tname] = TurbineType.new(
@@ -150,14 +150,14 @@ def read_layout(lname, ldict, algo_dict, ttype, verbosity=1):
     :group: input.windio
 
     """
-    if verbosity > 1:
+    if verbosity > 2:
         print(f"        Reading '{lname}'")
     cdict = Dict(ldict["coordinates"], name="coordinates")
     farm = algo_dict["farm"]
     for xy in zip(cdict["x"], cdict["y"]):
         farm.add_turbine(
             Turbine(xy=np.array(xy), turbine_models=[ttype]),
-            verbosity=0,
+            verbosity=verbosity-3,
         )
-    if verbosity > 1:
+    if verbosity > 2:
         print(f"          Added {farm.n_turbines} wio_trbns of type '{ttype}'")