+++ foxes/output/__init__.py
@@ -14,6 +14,7 @@ from .animation import Animator
 from .calc_points import PointCalculator
 from .slice_data import SliceData
 from .rotor_point_plots import RotorPointPlot
+from .state_turbine_table import StateTurbineTable
 
 from .flow_plots_2d import FlowPlots2D, SeqFlowAnimationPlugin
 from . import grids